Rename language::tests -> language::buffer_tests
